fixes and small renderer changes

+++ src/main/java/de/omnikryptec/render/objects/Sprite.java
@@ -6,24 +6,24 @@ import de.omnikryptec.render.batch.Batch2D;
 
 public abstract class Sprite implements RenderedObject {
     public static final RenderedObjectType TYPE = RenderedObjectType.of(Sprite.class);
-
-    private float layer;
-
+    
+    private int layer;
+    
     @Override
     public boolean isVisible(final FrustumIntersection frustum) {
         return true;
     }
-
-    public float getLayer() {
+    
+    public int getLayer() {
         return this.layer;
     }
-
-    public void setLayer(final float layer) {
+    
+    public void setLayer(final int layer) {
         this.layer = layer;
     }
-
+    
     public abstract void draw(Batch2D batch);
-
+    
     @Override
     public RenderedObjectType type() {
         return TYPE;
